requirements.txt
pandas==2.2.3
sodapy==2.2.0
PyYAML==6.0.2
pyarrow==19.0.1
scipy==1.15.2
scikit-learn==1.6.1
python-dotenv==1.1.0
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
nltk==3.9.1